requirements.txt
tensorflow[and-cuda]
matplotlib
ipykernel
pydot
pytest
pydantic
pyyaml
keras-cv